+++ Admin/CoreSetting/SettingsTab.php
@@ -11,6 +11,7 @@ namespace OxidEsales\Codeception\Admin\CoreSetting;
 
 use OxidEsales\Codeception\Module\Translation\Translator;
 use OxidEsales\Codeception\Page\Page;
+use OxidEsales\EshopCommunity\Tests\Codeception\AcceptanceTester;
 
 /**
  * Class SettingsTab
@@ -61,5 +62,35 @@ class SettingsTab extends Page
         $I->waitForElementVisible(['class' => 'yui-dt-data']);
 
         return new StartCategoryFrontendPopup($I);
+    }
+
+    public function openAdministration(): SettingsTab
+    {
+        /** @var AcceptanceTester $I */
+        $I = $this->user;
+        $I->selectEditFrame();
+        $I->click(Translator::translate('SHOP_OPTIONS_GROUP_ADMINISTRATION'));
+        $I->waitForPageLoad();
+        return $this;
+    }
+
+    public function setAdminFormat(string $format): SettingsTab
+    {
+        /** @var AcceptanceTester $I */
+        $I = $this->user;
+        $I->selectOption('confstrs[sLocalDateFormat]', $format);
+        $I->seeOptionIsSelected('confstrs[sLocalDateFormat]', $format);
+        $I->waitForPageLoad();
+        return $this;
+    }
+
+    public function save(): SettingsTab
+    {
+        /** @var AcceptanceTester $I */
+        $I = $this->user;
+        $I->selectEditFrame();
+        $I->click('save');
+        $I->waitForPageLoad();
+        return $this;
     }
 }