+++ CHANGELOG.md
@@ -19,12 +19,22 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Methods:
     - `OxidEsales\Codeception\Step\Basket::addProductToBasketAndOpen`
 
-## [2.2.0] - Unreleased
+## [2.2.0] - 2021-11-26
 
 ### Added
 - Trait:
-  - `Admin/Component/HeaderLinks`
-
+  - `Admin\Component\HeaderLinks`
+- Page Object:
+  - `Admin\CoreSetting\PerformanceTab`
+- Methods:
+  - `Admin\CoreSetting\SettingsTab`
+    - `openAdministration()`,
+    - `setAdminFormat()`,
+    - `save()`
+  - `Admin\Product\MainProductPage::save()`
+- Property:
+  - `Admin\Order\OrderList::$orderBillingLastNameInput`
+        
 ## [2.1.0] - 2021-07-09
 
 ### Added
@@ -197,7 +207,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - First version of the module introduced
 
-[3.0.0]: https://github.com/OXID-eSales/codeception-page-object/compare/v2.1.0..v3.0.0
+[3.0.0]: https://github.com/OXID-eSales/codeception-page-object/compare/v2.2.0..v3.0.0
+[2.2.0]: https://github.com/OXID-eSales/codeception-page-object/compare/v2.1.0..v2.2.0
 [2.1.0]: https://github.com/OXID-eSales/codeception-page-object/compare/v2.0.0..v2.1.0
 [2.0.0]: https://github.com/OXID-eSales/codeception-page-object/compare/v1.4.0..v2.0.0
 [1.4.0]: https://github.com/OXID-eSales/codeception-page-object/compare/v1.3.3..v1.4.0